closes #180 #193 #194 (#192)
* adding rentacar presentation

* adding rentacar presentation

* improve car interface

* changing int -> Integer and double -> Double

* adding cancel renting functionality

* closes #194 and closes #193

* polishing things up

* fixing when error occurs

* fix html tag

* fixing a couple of issues

* fixing potential NPE issue

* check null Integer and null Double

+++ car/src/test/java/pt/ulisboa/tecnico/softeng/car/domain/CarPersistenceTest.java
@@ -55,12 +55,12 @@ public class CarPersistenceTest {
 		for (Vehicle vehicle : rentACar.getVehicleSet()) {
 			if (vehicle instanceof Car) {
 				assertEquals(PLATE_CAR1, vehicle.getPlate());
-				assertEquals(10, vehicle.getKilometers());
+				assertEquals(10, vehicle.getKilometers().intValue());
 				assertEquals(10, vehicle.getPrice(), 0);
 			}
 			if (vehicle instanceof Motorcycle) {
 				assertEquals(PLATE_CAR2, vehicle.getPlate());
-				assertEquals(20, vehicle.getKilometers());
+				assertEquals(20, vehicle.getKilometers().intValue());
 				assertEquals(5, vehicle.getPrice(), 0);
 			}
 		}
